Adapt tournament.py and test to use the new schema.
The new schema stores wins in the participant table, and uses composite
primary keys.
When deleteding data from matchs we need to keep the participants table in
sync.
In the future, standing will actually show the number of points, instead of number of wins, since we want to add "tie" support to our program.
In this schema, a win is worth 3 points and a tie 1 point.